api/requirements.txt
fastapi
uvicorn
python-jose
passlib
python-dotenv